refactor(desktop): decompose note editor and settings components (#185)
